Add cancel button to import modal
