+++ renderer/components/ImportAccountModal/ImportAccountModal.tsx
@@ -1,4 +1,10 @@
 import {
+  Tabs,
+  TabList,
+  Tab,
+  TabPanels,
+  TabPanel,
+  Box,
   Modal,
   ModalOverlay,
   ModalContent,
@@ -6,10 +12,14 @@ import {
   Heading,
   useToast,
 } from "@chakra-ui/react";
-import { useCallback } from "react";
+import { useCallback, useMemo, useState } from "react";
 import { FormattedMessage, defineMessages, useIntl } from "react-intl";
 
-import { ImportAccount } from "../ImportAccount/ImportAccount";
+import { trpcReact } from "@/providers/TRPCProvider";
+
+import { EncodedKeyImport } from "./EncodedKeyImport";
+import { FileImport } from "./FileImport";
+import { MnemonicImport } from "./MnemonicImport";
 
 const messages = defineMessages({
   successTitle: {
@@ -19,6 +29,26 @@ const messages = defineMessages({
     defaultMessage: "Your account was successfully imported",
   },
 });
+
+function useTabsComponents() {
+  return useMemo(
+    () => [
+      {
+        label: <FormattedMessage defaultMessage="Mnemonic Phrase" />,
+        Component: MnemonicImport,
+      },
+      {
+        label: <FormattedMessage defaultMessage="Encoded Key" />,
+        Component: EncodedKeyImport,
+      },
+      {
+        label: <FormattedMessage defaultMessage="File" />,
+        Component: FileImport,
+      },
+    ],
+    [],
+  );
+}
 
 type Props = {
   isOpen: boolean;
@@ -31,6 +61,29 @@ export function ImportAccountModal({ isOpen, onClose }: Props) {
   }, [onClose]);
   const toast = useToast();
   const { formatMessage } = useIntl();
+
+  const [activeTab, setActiveTab] = useState(0);
+
+  const {
+    mutate: importAccount,
+    isLoading,
+    error,
+    reset,
+  } = trpcReact.importAccount.useMutation({
+    onSuccess: () => {
+      toast({
+        title: formatMessage(messages.successTitle),
+        description: formatMessage(messages.successDescription),
+        status: "success",
+        duration: 5000,
+        isClosable: true,
+      });
+      onClose();
+    },
+  });
+
+  const tabsComponents = useTabsComponents();
+
   return (
     <Modal isOpen={isOpen} onClose={handleClose}>
       <ModalOverlay />
@@ -39,18 +92,42 @@ export function ImportAccountModal({ isOpen, onClose }: Props) {
           <Heading fontSize="2xl" mb={4}>
             <FormattedMessage defaultMessage="Import Account" />
           </Heading>
-          <ImportAccount
-            onSuccess={() => {
-              toast({
-                title: formatMessage(messages.successTitle),
-                description: formatMessage(messages.successDescription),
-                status: "success",
-                duration: 5000,
-                isClosable: true,
-              });
-              onClose();
-            }}
-          />
+          <Box>
+            <Tabs
+              isLazy
+              index={activeTab}
+              onChange={(index) => {
+                reset();
+                setActiveTab(index);
+              }}
+            >
+              <TabList mb={8}>
+                {tabsComponents.map(({ label }, i) => (
+                  <Tab py={2} px={4} mr={4} key={i}>
+                    {label}
+                  </Tab>
+                ))}
+              </TabList>
+
+              <TabPanels>
+                {tabsComponents.map(({ Component }, i) => (
+                  <TabPanel p={0} key={i}>
+                    <Component
+                      isLoading={isLoading}
+                      error={error?.shape?.message}
+                      onImport={({ name, account }) => {
+                        importAccount({
+                          name,
+                          account,
+                        });
+                      }}
+                      onCancel={onClose}
+                    />
+                  </TabPanel>
+                ))}
+              </TabPanels>
+            </Tabs>
+          </Box>
         </ModalBody>
       </ModalContent>
     </Modal>
